Create README.md
The client-side component of this app is using REACT. A single-page responsive app that has rich interactions, several interface views, and routing along with a polished user experience. This will also support an existing server-side API by enabling user requests with a response with the server-side with several different interface views. 

The main purpose of this app was to present a fullstack application using REACT.